Fixed removal of "separate" parts This fixes issues involving removal or addition of parts with parent types considered "separate." (World, Hud, Elytra, etc.) This is done by running `AvatarRenderer.sortParts()` when it should be run.
Basic testing was done and it seems to work for every case I gave it.

Due to the nature of this bug, `:removeChild()` always has to sort parts again even if the child being removed is not
separate because that child could contain a child that *is* separate (and so on.)
If anyone has a better solution feel free to do whatever.